Fixed clicknupload_co (was _org)
